+++ packages/client/src/App.tsx
@@ -2,13 +2,13 @@ import './App.css';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { HomePage } from './pages/Home';
 import { ThemeProvider } from './theme/ThemeProvider';
-import { NavBar } from './components/NavigationBar';
+import { NavBar } from './components/NavigationBar.component';
 import { NewProject } from './pages/projects/NewProject';
 import { ProjectControl } from './pages/projects/ProjectControl';
 import { SuccessPage } from './pages/SuccessPage';
 import { NewStudy } from './pages/studies/NewStudy';
 import { ContributePage } from './pages/contribute/Contribute';
-import { TagView } from './components/TagView';
+import { TagView } from './components/TagView.component';
 import { StudyControl } from './pages/studies/StudyControl';
 import { ProjectAccess } from './pages/datasets/ProjectAccess';
 import { ProjectUserPermissions } from './pages/projects/ProjectUserPermissions';
@@ -22,7 +22,7 @@ import { AdminGuard } from './guards/AdminGuard';
 import { LogoutPage } from './pages/LogoutPage';
 import { CssBaseline, Box, styled } from '@mui/material';
 import { FC, ReactNode, useState } from 'react';
-import { SideBar } from './components/SideBar';
+import { SideBar } from './components/SideBar.component';
 import { ProjectProvider } from './context/ProjectContext';
 import { ApolloClient, ApolloProvider, InMemoryCache, concat, createHttpLink } from '@apollo/client';
 import { setContext } from '@apollo/client/link/context';
